Move toOperationId() to AbstractRubyCodegen
